LottoMake Test코드 수정
